backend/pyproject.toml
[tool.poetry]
name = "noteapp-backend"
version = "0.1.0"
description = "Notes app backend"
authors = ["Rafael Fragoso <rafaelfragosom@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}, {include = "apps"}]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
redis = "^5.0"
celery = "^5.3"
gunicorn = "^21.2"
drf-spectacular = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.7"
pytest-cov = "^4.1"
black = "^24.1"
ruff = "^0.2"
mypy = "^1.8"
watchdog = "^4.0"
ipdb = "^0.13"
django-extensions = "^3.2"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.tox
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*/migrations/*",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/settings/*",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
